+++ Assets/VRM10_Samples/VRM10Viewer/VRM10ViewerUI.cs
@@ -410,7 +410,7 @@ namespace UniVRM10.VRM10Viewer
                         m_loaded.Runtime.VrmAnimation = Motion;
                     }
 
-                    m_loaded.Runtime.IsSpringBoneEnabled = !m_pauseSpringBone.isOn;
+                    m_loaded.Runtime.SpringBone.IsSpringBoneEnabled = !m_pauseSpringBone.isOn;
                 }
             }
         }
